backend/services/lambda-service/requirements.txt
# Lambda runtime dependencies only (exclude dev/test dependencies)
fastapi[standard]==0.119.1
pydantic==2.9.0
python-dotenv==1.0.0
boto3==1.40.56
mangum==0.19.0
pymongo==4.6.1
pymongo[srv]==4.6.1
bcrypt==4.1.2
faker==30.8.2
